// frontend/generated/models/ProfileDto.ts
export interface ProfileDto {
    /**
     * 
     * @type {number}
     * @memberof ProfileDto
     */
    profileId?: number;
    /**
     * 
     * @type {AccountDto}
     * @memberof ProfileDto
     */
    account?: AccountDto;
    /**
     * 
     * @type {string}
     * @memberof ProfileDto
     */
    overview?: string;
    /**
     * 
     * @type {string}
     * @memberof ProfileDto
     */
    education?: string;
    /**
     * 
     * @type {string}
     * @memberof ProfileDto
     */
    phoneNumber?: string;
    /**
     * 
     * @type {string}
     * @memberof ProfileDto
     */
    language?: string;
    /**
     * 
     * @type {Array<ProfileSkillDto>}
     * @memberof ProfileDto
     */
    skills?: Array<ProfileSkillDto>;
    /**
     * 
     * @type {boolean}
     * @memberof ProfileDto
     */
    isVisible?: boolean;
    /**
     * 
     * @type {Date}
     * @memberof ProfileDto
     */
    createdAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof ProfileDto
     */
    updatedAt?: Date;
}